requirements.txt
wheel
pyparted
psutil
